💄 Separate out admin tasks
- cleanup to help us see what needs to be done to decouple the client & server
